refactor: extract enginePodBuilder, ModelConfig, and parseResourceProfile
- Extract enginePodBuilder type: func(Model, ModelConfig) *Pod
- Each engine's xxxPodForModel already matches this signature
- Resolve builder in getModelConfig() via resolveEnginePodBuilder()
- Store on ModelConfig.PodBuilder, eliminating engine switch in:
  - calculatePodPlan (was 4-way switch)
  - buildLeaderWorkerSet (was hardcoded vLLM check)
- Extract ModelConfig, LWSConfig, getModelConfig, lookupServerImage to
  model_config.go (was mixed into model_controller.go)
- Extract ResourceProfileRef and parseResourceProfile() for explicit
  resource profile parsing (was inline string splitting)
- Add json:"-" tag on PodBuilder to fix JSON comparison in tests
