+++ internal/modelcontroller/model_controller.go
@@ -21,7 +21,6 @@ import (
 	"errors"
 	"fmt"
 	"reflect"
-	"strconv"
 	"strings"
 	"time"
 
@@ -252,133 +251,6 @@ func (r *ModelReconciler) annotationsForModel(m *kubeaiv1.Model) map[string]stri
 	}
 
 	return ann
-}
-
-type ModelConfig struct {
-	config.CacheProfile
-	config.ResourceProfile
-	Image     string
-	Source    modelSource
-	LWSConfig *LWSConfig
-}
-
-// LWSConfig holds multi-node group configuration for LeaderWorkerSet.
-type LWSConfig struct {
-	TensorParallelSize   int
-	PipelineParallelSize int // Also used as LWS group size
-}
-
-func (r *ModelReconciler) getModelConfig(model *kubeaiv1.Model) (ModelConfig, error) {
-	var result ModelConfig
-
-	src, err := r.parseModelSource(model.Spec.URL)
-	if err != nil {
-		return result, fmt.Errorf("parsing model source: %w", err)
-	}
-	result.Source = src
-
-	if model.Spec.CacheProfile != "" {
-		cacheProfile, ok := r.CacheProfiles[model.Spec.CacheProfile]
-		if !ok {
-			return result, fmt.Errorf("cache profile not found: %q", model.Spec.CacheProfile)
-		}
-		result.CacheProfile = cacheProfile
-	}
-
-	split := strings.Split(model.Spec.ResourceProfile, ":")
-	if len(split) != 2 && len(split) != 3 {
-		return result, fmt.Errorf("invalid resource profile: %q, should match <name>:<count> or <name>:<tp>:<pp>, example: nvidia-gpu-l4:2 or nvidia-gpu-l4:2:3", model.Spec.ResourceProfile)
-	}
-	name := split[0]
-	multiple, err := strconv.Atoi(split[1])
-	if err != nil {
-		return result, fmt.Errorf("invalid multiple in resource profile: %q: %w", split[1], err)
-	}
-	if len(split) == 3 {
-		ppSize, err := strconv.Atoi(split[2])
-		if err != nil {
-			return result, fmt.Errorf("invalid pipeline-parallel size in resource profile: %q: %w", split[2], err)
-		}
-		if model.Spec.Engine != kubeaiv1.VLLMEngine {
-			return result, fmt.Errorf("multi-node (LWS) resource profiles only supported with VLLM engine, got %q", model.Spec.Engine)
-		}
-		result.LWSConfig = &LWSConfig{
-			TensorParallelSize:   multiple,
-			PipelineParallelSize: ppSize,
-		}
-	}
-
-	profile, ok := r.ResourceProfiles[name]
-	if !ok {
-		return result, fmt.Errorf("resource profile not found: %q", name)
-	}
-
-	requests := make(corev1.ResourceList)
-	for key, quantity := range profile.Requests {
-		q := quantity.DeepCopy()
-		q.Mul(int64(multiple))
-		requests[key] = q
-	}
-
-	limits := make(corev1.ResourceList)
-	for key, quantity := range profile.Limits {
-		q := quantity.DeepCopy()
-		q.Mul(int64(multiple))
-		limits[key] = q
-	}
-
-	result.ResourceProfile = profile
-	// Apply the multiplied requests and limits to the profile.
-	result.Requests = requests
-	result.Limits = limits
-
-	if model.Spec.EnvFrom != nil {
-		result.Source.modelSourcePodAdditions.envFrom = model.Spec.EnvFrom
-	}
-
-	image, err := r.lookupServerImage(model, profile)
-	if err != nil {
-		return result, fmt.Errorf("looking up server image: %w", err)
-	}
-	result.Image = image
-
-	return result, nil
-}
-
-func (r *ModelReconciler) lookupServerImage(model *kubeaiv1.Model, profile config.ResourceProfile) (string, error) {
-	if model.Spec.Image != "" {
-		return model.Spec.Image, nil
-	}
-
-	var serverImgs map[string]string
-	switch model.Spec.Engine {
-	case kubeaiv1.OLlamaEngine:
-		serverImgs = r.ModelServers.OLlama.Images
-	case kubeaiv1.FasterWhisperEngine:
-		serverImgs = r.ModelServers.FasterWhisper.Images
-	case kubeaiv1.InfinityEngine:
-		serverImgs = r.ModelServers.Infinity.Images
-	default:
-		serverImgs = r.ModelServers.VLLM.Images
-	}
-
-	// If no image name is provided for a profile, use the default image name.
-	const defaultImageName = "default"
-	imageName := defaultImageName
-	if profile.ImageName != "" {
-		imageName = profile.ImageName
-	}
-
-	if img, ok := serverImgs[imageName]; ok {
-		return img, nil
-	}
-
-	// If the specific profile image name does not exist, use the default image name.
-	if img, ok := serverImgs[defaultImageName]; ok {
-		return img, nil
-	} else {
-		return "", fmt.Errorf("missing default server image")
-	}
 }
 
 func (r *ModelReconciler) applyAutoscalingReplicaBounds(model *kubeaiv1.Model) bool {
